Fixed issue where tests were populated when entering a URL on the landing page
For those who are interested to know what the bug was, the `model` method
on routes only gets fired when navigating to a route where a model is not
passed in. Specifically, passing a model to {{link-to}}, Controller#transitionToRoute,
or Route#transitionTo will skip the `model` method.

Any route that needs to fetch more data should use the `afterModel` method.
If a promise is returned by `afterModel` Ember will wait until that promise
resolves before finalizing the transition.